+++ application/translation/nld.go
@@ -11,9 +11,9 @@ func (coreNL) Language() configuration.Language { return configuration.LangNL }
 func (coreNL) Translations() map[string]string {
 	return map[string]string{
 		// Domain validation
-		"RequiredFieldNotification":            "Verplicht veld.",
-		"SchemaViolationNotification":          "De inhoud van de verzoekbody komt niet overeen met het verwachte schema.",
-		"LimitExceededNotification":            "De gevraagde limiet overschrijdt het toegestane maximum.",
+		"RequiredFieldNotification":   "Verplicht veld.",
+		"SchemaViolationNotification": "De inhoud van de verzoekbody komt niet overeen met het verwachte schema.",
+		"LimitExceededNotification":   "De gevraagde limiet overschrijdt het toegestane maximum.",
 
 		// Domain entity
 		"UnableToInsertWithIDNotification":    "Kan een record met een bestaande primaire sleutel niet invoegen.",